feat: add collaborative movie submission and ownership permissions

+++ movies/views/favorite_views.py
@@ -1,19 +1,35 @@
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
+from rest_framework.permissions import IsAuthenticated
 from django.shortcuts import get_object_or_404
 
 from movies.serializers import FavoriteSerializer
 from movies.models import Favorite
+from movies.permissions.favorite_permissions import IsOwnerFavorite
 
 class FavoriteView(APIView):
+
+    def get_permissions(self):
+
+        if self.request.method in ["GET", "POST"]:
+            return [IsAuthenticated()]
+
+        elif self.request.method == "DELETE":
+            return [
+                IsAuthenticated(),
+                IsOwnerFavorite()
+            ]
+
+        return super().get_permissions()
+    
     
     def post(self, request):
         
         serializer = FavoriteSerializer(data=request.data)
         
         if serializer.is_valid():
-            serializer.save()
+            serializer.save(user=request.user)
             
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         
@@ -22,7 +38,7 @@ class FavoriteView(APIView):
     
     def get(self, request):
         
-        favorites = Favorite.objects.all()
+        favorites = Favorite.objects.filter(user=request.user)
         
         serializer = FavoriteSerializer(
             favorites,
@@ -34,8 +50,20 @@ class FavoriteView(APIView):
     
     def delete(self, request, pk=None):
         
-        favorite = get_object_or_404(Favorite, pk=pk)
+        favorite = get_object_or_404(
+            Favorite,
+            pk=pk,
+            )
         
+        self.check_object_permissions(
+            request,
+            favorite
+        )
+            
         favorite.delete()
         
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response(
+            status=status.HTTP_204_NO_CONTENT
+            )
+        
+
